Support matching by value
This adds support for matching labels and annotations with specific
values. For example, a label matcher of the format
`app=kpack-controller` will only match a pod with the label `app` set to
`kpack-controller`. This allows for more precise matching.

In the future we could consider allowing multiple, comma-separated
values, for now having multiple entries for multiple values should do.

Signed-off-by: Giuseppe Capizzi <gcapizzi@vmware.com>